pyproject.toml
[tool.poetry]
name = "my-blog"
version = "0.1.0"
description = ""
authors = ["GamBit1986 <mazepa86@gmail.com>"]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
flask-login = "^0.6.2"
flask-migrate = "^4.0.4"
python-dotenv = "^1.0.0"
poetry = "^1.4.2"
psycopg2-binary = "^2.9.6"
wtforms = "^3.0.1"
email-validator = "^1.3.1"
flask-wtf = "^1.1.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
